Feature/sprint4

+++ app/Http/Controllers/Accounts/BankPaymentController.php
@@ -46,7 +46,7 @@ class BankPaymentController extends Controller
 
             $dataSql = Voucher::where('type',self::Constants()['type'])->distinct()->orderby('date','desc');
 
-            $allData = $dataSql->get(['voucher_id','voucher_no','date','remarks']);
+            $allData = $dataSql->get(['voucher_id','voucher_no','date','posted','remarks']);
 
             $recordsTotal = count($allData);
             $recordsFiltered = count($allData);
@@ -61,6 +61,7 @@ class BankPaymentController extends Controller
             }
             $entries = [];
             foreach ($allData as $row) {
+                $posted = $this->getPostedTitle()[$row->posted];
                 $urlEdit = route('accounts.bank-payment.edit',$row->voucher_id);
                 $urlDel = route('accounts.bank-payment.destroy',$row->voucher_id);
 
@@ -77,10 +78,10 @@ class BankPaymentController extends Controller
                     $actions .= '<a href="' . $urlEdit . '" class="item-edit"><i data-feather="edit"></i></a>';
                 }
                 $actions .= '</div>'; //end main div
-
                 $entries[] = [
                     $row->date,
                     $row->voucher_no,
+                    '<div class="text-center"><span class="badge rounded-pill ' . $posted['class'] . '">' . $posted['title'] . '</span></div>',
                     $row->remarks,
                     $actions,
                 ];
@@ -108,7 +109,7 @@ class BankPaymentController extends Controller
         $data['title'] = self::Constants()['title'];
         $data['list_url'] = self::Constants()['list_url'];
         $data['permission'] = self::Constants()['create'];
-        $max = Voucher::where('type',self::Constants()['type'])->max('voucher_no');
+        $max = Voucher::withTrashed()->where('type',self::Constants()['type'])->max('voucher_no');
         $data['voucher_no'] = self::documentCode(self::Constants()['type'],$max);
       //  $data['payment_mode'] = PaymentMode::where('status',1)->get();
         return view('accounts.bank_payment.create', compact('data'));
@@ -153,10 +154,10 @@ class BankPaymentController extends Controller
         DB::beginTransaction();
         try {
 //dd("sef");
-            $max = Voucher::where('type',self::Constants()['type'])->max('voucher_no');
+            $max = Voucher::withTrashed()->where('type',self::Constants()['type'])->max('voucher_no');
             $voucher_no = self::documentCode(self::Constants()['type'],$max);
             $voucher_id = self::uuid();
-
+            $posted = $request->current_action_id == 'post'?1:0;
             $sr = 1;
             foreach ($request->pd as $pd){
                 $account = ChartOfAccount::where('id',$pd['chart_id'])->first();
@@ -180,6 +181,7 @@ class BankPaymentController extends Controller
                         'company_id' => auth()->user()->company_id,
                         'project_id' => auth()->user()->project_id,
                         'user_id' => auth()->user()->id,
+                        'posted' => $posted,
                     ]);
                     $sr = $sr + 1;
                 }
@@ -229,7 +231,11 @@ class BankPaymentController extends Controller
         }
 
         $data['view'] = false;
-        if(isset($request->view)){
+        $data['posted'] = false;
+        if($data['current']->posted == 1){
+            $data['posted'] = true;
+        }
+        if(isset($request->view) || $data['current']->posted == 1){
             $data['view'] = true;
             $data['permission'] = self::Constants()['view'];
             $data['permission_edit'] = self::Constants()['edit'];
@@ -281,10 +287,15 @@ class BankPaymentController extends Controller
         try {
 
             $firstVoucher = Voucher::where('type',self::Constants()['type'])->where('voucher_id',$id)->first();
+            if($firstVoucher->posted == 1){
+                return $this->jsonErrorResponse($data, 'This voucher have been already posted');
+            }
+
             $voucher_no = $firstVoucher->voucher_no;
             $voucher_id = $id;
             DB::select("delete FROM `vouchers` where voucher_id = '$voucher_id'");
 
+            $posted = $request->current_action_id == 'post'?1:0;
             $sr = 1;
             foreach ($request->pd as $pd){
                 $account = ChartOfAccount::where('id',$pd['chart_id'])->first();
@@ -308,6 +319,7 @@ class BankPaymentController extends Controller
                         'company_id' => auth()->user()->company_id,
                         'project_id' => auth()->user()->project_id,
                         'user_id' => auth()->user()->id,
+                        'posted' => $posted,
                     ]);
                     $sr = $sr + 1;
                 }
@@ -331,6 +343,17 @@ class BankPaymentController extends Controller
      */
     public function destroy($id)
     {
-        //
+        $data = [];
+        DB::beginTransaction();
+        try{
+
+            Voucher::where('voucher_id',$id)->delete();
+
+        }catch (Exception $e) {
+            DB::rollback();
+            return $this->jsonErrorResponse($data, $e->getMessage(), 200);
+        }
+        DB::commit();
+        return $this->jsonSuccessResponse($data, 'Successfully deleted', 200);
     }
 }
